makes separate traits for kfolds, monte carlo, and progressive
